fix: remove generic from Buffer
Buffers are defined as `Buffer<T>`, but the generic isn't used with the
actual underlying OpenCL buffer (which is set to `u8`). It was always
the case, since the inception of this library. Hence, I don't see a
reason to keep the generic. Especially since it is misleading, you could
define `Buffer<16>`, but the underlying OpenCL buffer would still be `u8`.