feat: better unique identifier support
Using the UUID as unique identifier is preferred, though it is not always
available. Hence there is a fallback to the PCI ID, which is always
available on all supported devices.

BREAKING CHANGE: Bus ID is replaced with PCI ID

Prior to this change a so-called "Bus ID" was used. Instead now use a
"PCI ID", which should provide more uniqueness.

Also the `GPUSelector` is replaced by with additional functions on
`Device` which provide the same functionality.