feat: change kernel parameter handling
BREAKING CHANGE: Kernel creation takes different parameter values now.

Defining the global work size now follows the CUDA model and no longer
the OpenCL model. In OpenCL the global work size is the total number
of threads. In CUDA it is the number of thread groups of local work
size.

This change is made, so that the same kernel creation code can be run
indepent of whether OpenCL or CUDA is used.