feat: add CUDA support
This library is now an abstraction for OpenCL and CUDA. It can be compiled
with support for either or both, via the feature flags `opencl` and `cuda`.

There is now a `Device`, which can point to an OpenCL and/or CUDA device.

You should be able to execute OpenCL and CUDA kernels with the same code
without modifications. You would pass in two closures with the same function
body into `Program::run()`.

To create two closures with the correct signature but the same body, you
can use the `define_closures()` helper macro.

So your code would look like this:

    use rust_gpu_tools::{cuda, define_closures, opencl, GPUError, Program};

    let closures = define_closures!(|program: &opencl::Program | &cuda::Program|
     -> Result<Vec<u8>, GPUError> {
        let input = program.create_buffer::<u8>(128)?;
        let output = program.create_buffer::<u8>(128)?;
        let kernel = program.create_kernel(…)?;
        kernel.arg(&input).arg(&output).run()?;
        let mut out = vec![0u8; 128];
        program.read_into_buffer(&output, 0, &mut out)?;
        Ok(out)
    });

    let program = Program::Cuda(cuda::Program::from_cuda(…)?;
    let results = program.run(closures)?;